// src/shared/api/models/ProjectParentTaskDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { MemberDto } from "./MemberDto";
import type { ProjectTaskDto } from "./ProjectTaskDto";

export type ProjectParentTaskDto = {
  /**
   * @description ID родительской задачи
   * @type number
   */
  id: number;
  /**
   * @description Наименование родительской задачи
   * @type string
   */
  name: string;
  /**
   * @description Описание родительской задачи
   * @type string
   */
  description?: string | null;
  /**
   * @description Автор задачи
   */
  author: MemberDto;
  /**
   * @description Дедлайн родительской задачи
   * @type string
   */
  deadline_date?: string | null;
  /**
   * @description Дата завершения родительской задачи
   * @type string
   */
  done_date?: string | null;
  /**
   * @description Дочерние задачи
   * @type array
   */
  project_tasks?: ProjectTaskDto[] | null;
};