// src/shared/api/models/UpdateProjectParentTaskBody.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type UpdateProjectParentTaskBody = {
  /**
   * @description Наименование родительской задачи
   * @minLength 1
   * @type string
   */
  name: string
  /**
   * @description Описание родительской задачи
   * @type string
   */
  description?: string | null
  /**
   * @description Дедлайн родительской задачи
   * @type string
   */
  deadline_date?: string | null
  /**
   * @description Дата завершения задачи
   * @type string
   */
  done_date?: string | null
  /**
   * @description Массив ID задач
   * @type array
   */
  project_task_ids?: number[] | null
}